+++ src/App.tsx
@@ -1,11 +1,11 @@
 import React, { useEffect, useState } from 'react';
 import './App.scss';
 import { BrowserRouter as Router, Route, Routes } from 'react-router-dom';
-import { AtpNavbar } from './components/navbar-atp/AtpNavbar';
+import { AtpNavbar } from './components/layout/navbar-atp/AtpNavbar';
 import { AtpArticlePage } from './pages/article-page/AtpArticlePage';
 import Home from './pages/home-page/Home';
 import AtpNotFound from './pages/page-not-found/AtpNotFound';
-import AtpFooter from './components/footer/AtpFooter';
+import AtpFooter from './components/layout/footer/AtpFooter';
 import { ARTICLES_DATA } from './assets/dummy-data/articlesData';
 import AtpShoppingCart from './pages/shopping-cart/AtpShoppingCart';
 import { Article } from './models/article.model';
