+++ src/components/article/AtpArticle.tsx
@@ -1,13 +1,13 @@
 import React, { FC, useState } from 'react';
 import './AtpArticle.scss';
-import AtpLink from '../link-atp/AtpLink';
-import AtpText from '../../components/text/AtpText';
-import AtpList from '../../components/list/AtpList';
-import AtpButton from '../../components/button/AtpButton';
+import AtpLink from '../common/link-atp/AtpLink';
+import AtpText from '../common/text/AtpText';
+import AtpList from '../common/list/AtpList';
+import AtpButton from '../common/button/AtpButton';
 import AtpPrice from '../../components/price/AtpPrice';
-import AtpSelect from '../../components/select/AtpSelect';
-import AtpCarousel from '../../components/carousel/AtpCarousel';
-import AtpModal from '../../components/modal/AtpModal';
+import AtpSelect from '../common/select/AtpSelect';
+import AtpCarousel from '../common/carousel/AtpCarousel';
+import AtpModal from '../common/modal/AtpModal';
 import { strings } from '../../constants/strings';
 import { HiOutlineHeart } from 'react-icons/hi';
 import { Article } from '../../models/article.model';
