+++ src/components/common/button/AtpButton.scss
@@ -1,4 +1,4 @@
-@import '../../assets/styles/main.scss';
+@import '../../../assets/styles/main.scss';
 
 .atp-button {
   border: $border-classic;
