+++ src/components/common/carousel/AtpCarousel.scss
@@ -1,4 +1,4 @@
-@import '../../assets/styles/main.scss';
+@import '../../../assets/styles/main.scss';
 
 .atp-carousel {
   position: relative;
