+++ src/components/common/link-atp/AtpLink.scss
@@ -1,4 +1,4 @@
-@import '../../assets/styles/main.scss';
+@import '../../../assets/styles/main.scss';
 
 .atp-link {
   color: $color-black;
