+++ src/components/common/list/AtpList.scss
@@ -1,4 +1,4 @@
-@import '../../assets/styles/main.scss';
+@import '../../../assets/styles/main.scss';
 
 .atp-list {
   font-size: $font-size-base;
