+++ src/components/common/modal/AtpModal.scss
@@ -1,4 +1,4 @@
-@import '../../assets/styles/main.scss';
+@import '../../../assets/styles/main.scss';
 
 .close-modal {
   margin-top: $spacing-md;
