+++ src/components/common/select/AtpSelect.scss
@@ -1,0 +1,11 @@
+@import '../../../assets/styles/main.scss';
+
+.atp-selector {
+  text-transform: uppercase;
+  appearance: none;
+  border-radius: 0;
+  font-weight: 500;
+  background: transparent url('../../../assets/images/icons/dropdow-arrow-2xmargin-right.png')
+    center right no-repeat;
+  background-size: $spacing-xl;
+}
