+++ src/components/common/text/AtpText.scss
@@ -1,0 +1,8 @@
+@import '../../../assets/styles/main.scss';
+
+.atp-text {
+  color: $color-black;
+  font-size: $font-size-base;
+  margin-block-start: 0;
+  margin-bottom: $spacing-xl;
+}
