+++ src/components/layout/footer/AtpFooter.scss
@@ -1,4 +1,4 @@
-@import '../../assets/styles/main.scss';
+@import '../../../assets/styles/main.scss';
 
 .atp-footer {
   margin-top: $spacing-xl;
