+++ src/components/layout/navbar-atp/AtpNavbar.scss
@@ -1,4 +1,4 @@
-@import '../../assets/styles/main.scss';
+@import '../../../assets/styles/main.scss';
 
 .atp-navbar {
   position: fixed;
