--- src/components/text/AtpText.scss
@@ -1,8 +1,0 @@
-@import '../../assets/styles/main.scss';
-
-.atp-text {
-    color: $color-black;
-    font-size: $font-size-base;
-    margin-block-start: 0;
-    margin-bottom: $spacing-xl;
-}
