+++ src/pages/shopping-cart/AtpShoppingCart.tsx
@@ -1,7 +1,7 @@
 import React, { FC } from 'react';
-import AtpButton from '../../components/button/AtpButton';
-import AtpLink from '../../components/link-atp/AtpLink';
-import AtpText from '../../components/text/AtpText';
+import AtpButton from '../../components/common/button/AtpButton';
+import AtpLink from '../../components/common/link-atp/AtpLink';
+import AtpText from '../../components/common/text/AtpText';
 import { strings } from '../../constants/strings';
 import { CartItem } from '../../models/cartItem.model';
 import { calculateTotal, cart, removeFromCart } from '../../utilities/shoppingCart';
